perf(zkevm): compare jump-destination bytes at 64 bits (#13860)
* perf(zkevm): compare jump-destination bytes at 64 bits

The guest's byte scan held the opcode and its two thresholds in `int`, so the
JIT canonicalised both operands before each compare. The per-byte core ran six
instructions and four were sign extensions: two re-extending a byte `lb` had
already sign-extended, two re-extending thresholds that never change inside the
loop.

Holding all three as `nint` removes them. Block 26055159 on ziskemu: 159,316,645
-> 154,681,474 steps (-2.91%) and 18,755,349,025 -> 18,420,049,084 cost (-1.79%),
output byte-identical.

* Apply suggestion from @LukaszRozmej

+++ src/Nethermind/Nethermind.Evm/CodeAnalysis/JumpDestinationAnalyzer.zkevm.cs
@@ -43,8 +43,8 @@ public sealed partial class JumpDestinationAnalyzer
         long currentFlags = 0;
         nuint flagsPosition = 0;
         ref int thresholds = ref MemoryMarshal.GetArrayDataReference(_byteScanThresholds);
-        int jumpDest = thresholds;
-        int push1 = Unsafe.Add(ref thresholds, 1);
+        nint jumpDest = thresholds;
+        nint push1 = Unsafe.Add(ref thresholds, 1);
         // The PUSH skip below steps up to 32 bytes past the end of the code, so the walk pins and moves
         // an unmanaged pointer: the same overshoot on a `ref byte` is a managed pointer outside its
         // object, which a relocating GC may adjust wrongly even though it is only ever compared - and
@@ -58,7 +58,7 @@ public sealed partial class JumpDestinationAnalyzer
             {
                 // Sign extension folds everything above PUSH32 below JUMPDEST, so one signed comparison
                 // covers the whole [JUMPDEST, PUSH32] window that the rebase-and-range-test needed two for.
-                int op = (sbyte)*position;
+                nint op = (sbyte)*position;
                 if (op >= jumpDest)
                 {
                     if (op >= push1)
